FrontBackend/requirements.txt
fastapi
uvicorn
google-generativeai
reportlab
python-dotenv
psycopg2-binary
pydantic